Move config back into config project
